requirements.txt
opencv-python==4.0.1.24
Pillow==6.0.0